requirements.txt
requests-tor>=1.1.0
beautifulsoup4>=4.9.3
tqdm>=4.65.0
urllib3>=2.0.0
psycopg2-binary>=2.9.3
logging>=0.5.1.2
multiprocessing>=2.6.2
uuid>=1.30
hashlib>=20081119
xmltodict>=0.13.0
json>=2.0.9
xml.etree.ElementTree
xml.dom.minidom
